mysite/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-crispy-forms==1.14.0
django-taggit==3.0.0
mysqlclient==2.1.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
